Make all formatting tests always use default components and default namespaces
